Cambios JSX y login
