+++ src/components/IconLocation.jsx
@@ -1,0 +1,39 @@
+import { Icon } from 'leaflet'
+export const iconLocation = new Icon({
+  iconUrl: 'https://cdn-icons-png.flaticon.com/512/5632/5632722.png',
+  iconAnchor: null,
+  shadowUrl: null,
+  shadowSize: null,
+  shadowAnchor: null,
+  iconSize: [35, 35],
+  className: 'leaflet-venue-icon'
+})
+export const iconMarkerGreen = new Icon({
+  iconUrl: 'https://cdn-icons-png.flaticon.com/512/7946/7946073.png',
+  iconAnchor: null,
+  shadowUrl: null,
+  shadowSize: null,
+  shadowAnchor: null,
+  iconSize: [35, 35],
+  className: 'leaflet-venue-icon'
+})
+
+export const iconMarkerBlue = new Icon({
+  iconUrl: 'locationBlue.png',
+  iconAnchor: null,
+  shadowUrl: null,
+  shadowSize: null,
+  shadowAnchor: null,
+  iconSize: [35, 35],
+  className: 'leaflet-venue-icon'
+})
+
+export const iconMarkerMuseo = new Icon({
+  iconUrl: 'locationMuseo.png',
+  iconAnchor: null,
+  shadowUrl: null,
+  shadowSize: null,
+  shadowAnchor: null,
+  iconSize: [35, 35],
+  className: 'leaflet-venue-icon'
+})
