+++ src/index.css
@@ -93,69 +93,7 @@ code {
 }
 
 @import url('https://fonts.googleapis.com/css?family=Roboto');
- .auth__header {
-	 padding: 13vh 1rem calc(11vh + 35px);
-	 display: flex;
-	 align-items: center;
-	 justify-content: center;
-	 background-color: #f0f0f0;
-	 background-image: linear-gradient(#3280e4, #584dc3);
-	 background-size: cover;
-	 background-position: center center;
-	 position: relative;
-	 box-shadow: 0 3px 5px rgba(0, 0, 0, 0.3);
-}
- .auth__header:before {
-	 content: "";
-	 position: absolute;
-	 left: 0;
-	 top: 0;
-	 right: 0;
-	 bottom: 0;
-}
- .auth__logo {
-	 position: relative;
-	 z-index: 18;
-	 background: #fff;
-	 padding: 10px;
-	 border-radius: 50%;
-	 box-shadow: 0 2px 7px rgba(0, 0, 0, 0.15);
-	 overflow: hidden;
-}
- .auth__body {
-	 padding-bottom: 2rem;
-}
- .auth__form {
-	 min-width: 280px;
-	 max-width: 340px;
-	 margin: auto;
-	 margin-top: -40px;
-	 padding: 0 10px;
-	 position: relative;
-	 z-index: 9;
-}
- .auth__form_body {
-	 padding: 0.7rem 1.5rem 35px;
-	 border-radius: 0.5rem;
-	 background: #fff;
-	 border: 1px solid #eee;
-	 box-shadow: 0 4px 10px rgba(0, 0, 0, 0.15);
-}
- .auth__form_title {
-	 font-size: 1.3rem;
-	 text-align: center;
-	 text-transform: uppercase;
-	 margin-bottom: 1.2rem;
-}
- .auth__form_actions {
-	 text-align: center;
-	 padding: 0 2rem;
-	 margin-top: -25px;
-}
- .auth__form_actions .btn {
-	 border-radius: 30px;
-	 box-shadow: 0 2px 12px rgba(50, 128, 228, 0.5);
-}
+
 .backLoader {
   width: 100%;
   height: 100%;
