+++ src/screens/FormFirebase.jsx
@@ -2,6 +2,7 @@
 import { useState } from 'react'
 import { useAuth } from '../context/authContext'
 import { useNavigate } from 'react-router-dom'
+import bgImage from '../assets/bg.jpg'
 
 function FormsFirebase () {
   const auth = useAuth()
@@ -10,56 +11,53 @@ function FormsFirebase () {
   const [password, setPassword] = useState('')
   const handleLogin = (e) => {
     e.preventDefault()
-    auth.login(email, password)
-    navigate('/administration')
+    if (email != null && password != null && password !== '' && email !== '') {
+      auth.login(email, password).then(() => { navigate('/administration') }).catch(() => { alert('Couldn\'t login ' + email) })
+    }
   }
   const handleLogout = (e) => {
     e.preventDefault()
     auth.logout()
   }
-  const usuario = () => {
-    console.log(auth.user.email)
-  }
 
   return (
-    <div className='auth'>
-      <div className='auth__header'>
-        <div className='auth__logo'>
-          <img height='90' src='https://d2eip9sf3oo6c2.cloudfront.net/series/square_covers/000/000/083/full/EGH_VueJS_Final.png' alt='' />
-        </div>
-      </div>
-      <div className='auth__body'>
-        <form className='auth__form' autoComplete='off'>
-          <div className='auth__form_body'>
-            <h3 className='auth__form_title'>Sign in</h3>
-            <div>
-              <div className='form-group'>
-                <label className='text-uppercase small'>Email</label>
-                <input type='email' className='form-control' onChange={(e) => setEmail(e.target.value)} placeholder='Enter email' />
+    <div>
+      <meta charSet='UTF-8' />
+      <meta name='viewport' content='width=device-width, initial-scale=1.0' />
+      {/* =============== REMIXICONS =============== */}
+      {/* =============== CSS =============== */}
+      <div className='login'>
+        <img src={bgImage} alt='login' className='login__img' />
+        <form action className='login__form'>
+          <h1 className='login__title'>Bienvenido a GeoGuía</h1>
+          <div className='login__content'>
+            <div className='login__box'>
+              <i className='ri-user-3-line login__icon' />
+              <div className='login__box-input'>
+                <input type='email' required className='login__input' placeholder=' ' onChange={(e) => setEmail(e.target.value)} />
+                <label htmlFor className='login__label'>CorreoElectrónico</label>
               </div>
-              <div className='form-group'>
-                <label className='text-uppercase small'>Password</label>
-                <input type='password' onChange={(e) => setPassword(e.target.value)} className='form-control' placeholder='Password' />
+            </div>
+            <div className='login__box'>
+              <i className='ri-lock-2-line login__icon' />
+              <div className='login__box-input'>
+                <input type='password' required className='login__input' id='login-pass' placeholder=' ' onChange={(e) => setPassword(e.target.value)} />
+                <label htmlFor className='login__label'>Contraseña</label>
+                <i className='ri-eye-off-line login__eye' id='login-eye' />
               </div>
             </div>
           </div>
-          <div className='auth__form_actions'>
-            <button className='btn btn-primary btn-lg btn-block' onClick={(e) => handleLogin(e)}>
-              LOGIN
-            </button>
-            <div className='mt-2'>
-              <a href='#' className='small text-uppercase' onClick={() => usuario()}>
-                Forgot password
-              </a>
+          <div className='login__check'>
+            <div className='login__check-group'>
+              <input type='checkbox' className='login__check-input' />
+              <label htmlFor className='login__check-label'>Remember me</label>
             </div>
-            <div className='mt-2'>
-              <button className='btn btn-primary btn-lg btn-block' onClick={(e) => handleLogout(e)}>
-                logout
-              </button>
-            </div>
+            <a href='#' className='login__forgot' onClick={(e) => handleLogout(e)}>Forgot Password?</a>
           </div>
+          <button className='login__button' onClick={(e) => handleLogin(e)}>Login</button>
         </form>
       </div>
+      {/* =============== MAIN JS =============== */}
     </div>
   )
 }
